Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,23 +1,23 @@
 #include <stdio.h>
 int main() {
-    float a,b;
+    double a,b;
     char c;
-    scanf("%f %f %c",&a,&b,&c);
+    scanf("%lf %lf %c",&a,&b,&c);
     switch(c){
     case '+':
-    printf("%d",a+b);
+    printf("%lf",a+b);
     break;
 
     case '-':
-    printf("%d",a-b);
+    printf("%lf",a-b);
     break;
 
     case '*':
-    printf("%d",a*b);
+    printf("%lf",a*b);
     break;
 
     case '/':
-    printf("%d",a/b);
+    printf("%lf",a/b);
     break;
 
     default:
